[refact]: extract abstract connection
