+++ firefly/src/main/java/com/firefly/codec/http2/stream/AbstractHTTPConnection.java
@@ -1,5 +1,6 @@
 package com.firefly.codec.http2.stream;
 
+import com.firefly.codec.common.AbstractConnection;
 import com.firefly.codec.http2.model.HttpVersion;
 import com.firefly.net.ByteBufferArrayOutputEntry;
 import com.firefly.net.ByteBufferOutputEntry;
@@ -7,24 +8,18 @@ import com.firefly.net.SecureSession;
 import com.firefly.net.Session;
 import com.firefly.utils.function.Action1;
 import com.firefly.utils.function.Action2;
-import com.firefly.utils.io.IO;
 
 import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.util.Optional;
 
-abstract public class AbstractHTTPConnection implements HTTPConnection {
+abstract public class AbstractHTTPConnection extends AbstractConnection implements HTTPConnection {
 
-    protected final SecureSession secureSession;
-    protected final Session tcpSession;
     protected final HttpVersion httpVersion;
     protected volatile Object attachment;
     protected Action1<HTTPConnection> closedListener;
     protected Action2<HTTPConnection, Throwable> exceptionListener;
 
     public AbstractHTTPConnection(SecureSession secureSession, Session tcpSession, HttpVersion httpVersion) {
-        this.secureSession = secureSession;
-        this.tcpSession = tcpSession;
+        super(secureSession, tcpSession);
         this.httpVersion = httpVersion;
     }
 
@@ -41,29 +36,6 @@ abstract public class AbstractHTTPConnection implements HTTPConnection {
     @Override
     public HttpVersion getHttpVersion() {
         return httpVersion;
-    }
-
-    @Override
-    public boolean isClosed() {
-        return tcpSession.isClosed();
-    }
-
-    @Override
-    public boolean isOpen() {
-        return tcpSession.isOpen();
-    }
-
-    @Override
-    public void close() {
-        Optional.ofNullable(secureSession)
-                .filter(SecureSession::isOpen)
-                .ifPresent(IO::close);
-
-        Optional.ofNullable(tcpSession)
-                .filter(Session::isOpen)
-                .ifPresent(Session::close);
-
-        attachment = null;
     }
 
     public void writeEncryptMessage(Object message) throws IOException {
@@ -83,71 +55,6 @@ abstract public class AbstractHTTPConnection implements HTTPConnection {
     @Override
     public boolean isEncrypted() {
         return secureSession != null;
-    }
-
-    @Override
-    public int getSessionId() {
-        return tcpSession.getSessionId();
-    }
-
-    @Override
-    public long getOpenTime() {
-        return tcpSession.getOpenTime();
-    }
-
-    @Override
-    public long getCloseTime() {
-        return tcpSession.getCloseTime();
-    }
-
-    @Override
-    public long getDuration() {
-        return tcpSession.getDuration();
-    }
-
-    @Override
-    public long getLastReadTime() {
-        return tcpSession.getLastReadTime();
-    }
-
-    @Override
-    public long getLastWrittenTime() {
-        return tcpSession.getLastWrittenTime();
-    }
-
-    @Override
-    public long getLastActiveTime() {
-        return tcpSession.getLastActiveTime();
-    }
-
-    @Override
-    public long getReadBytes() {
-        return tcpSession.getReadBytes();
-    }
-
-    @Override
-    public long getWrittenBytes() {
-        return tcpSession.getWrittenBytes();
-    }
-
-    @Override
-    public long getIdleTimeout() {
-        return tcpSession.getIdleTimeout();
-    }
-
-    @Override
-    public long getMaxIdleTimeout() {
-        return tcpSession.getMaxIdleTimeout();
-    }
-
-    @Override
-    public InetSocketAddress getLocalAddress() {
-        return tcpSession.getLocalAddress();
-    }
-
-    @Override
-    public InetSocketAddress getRemoteAddress() {
-        return tcpSession.getRemoteAddress();
     }
 
     @Override
